+++ src/br/arthur/entities/Group.java
@@ -19,7 +19,7 @@ public class Group {
 	@Id
 	@Column(name="group_id")
 	@GeneratedValue
-	private short id;
+	private int id;
 	
 	@Column(name="name", nullable=false, unique=true)
 	private String name;
@@ -79,7 +79,7 @@ public class Group {
 		return true;
 	}
 
-	public short getId() {
+	public int getId() {
 		return id;
 	}
